Added some fixes according to Cr comments
