--- integrations/jira/utils.py
@@ -1,6 +1,0 @@
-from enum import StrEnum
-
-
-class ObjectKind(StrEnum):
-    PROJECT = "project"
-    ISSUE = "issue"
